added CommandsErrors go away from throwing exceptions

+++ src/core/commands/RemoveAllProductsCommand.java
@@ -5,21 +5,28 @@ import core.db.dao.ProductDAO;
 public class RemoveAllProductsCommand extends BaseCommandImp implements Command {
 
 	@Override
-	protected boolean isSpecificParamsIsValid() {
+	protected boolean isSpecificParamsValuesValid() {
 		return true;
 	}
 	
 	@Override
-	public void execute() throws Exception {
+	public boolean execute() throws Exception {
+		boolean isQueryOK = false;
 		String paramsList[] = {"password"};
-		checkParams(paramsList);			
-		
-		int rowsCount = ProductDAO.removeAllProducts(params.get("password"));
-		if (0==rowsCount) {
-			System.out.println("There is nothing to delete");
-			if (ProductDAO.isProductsExists())
-				System.out.printf("Products wasn't delete");
+		if (isParamsValid(paramsList)) {
+			int rowsCount = ProductDAO.removeAllProducts(params.get("password"));
+			if (-1 != rowsCount)
+				isQueryOK = true;
+			if (0 == rowsCount) {
+				System.out.println("There is nothing to delete");
+				if (ProductDAO.isProductsExists()) {
+					System.out.println("Products wasn't delete");
+					isQueryOK = false;
+				}
+			}
 		}
+		return isQueryOK;
+
 	}
 }
