+++ src/core/commands/RemoveProductCommand.java
@@ -5,7 +5,7 @@ import core.db.dao.ProductDAO;
 public class RemoveProductCommand extends BaseCommandImp implements Command {
 
 	@Override
-	protected boolean isSpecificParamsIsValid() {
+	protected boolean isSpecificParamsValuesValid() {
 		try {
 			Integer.parseInt(params.get("id"));
 		} catch (Exception e) {
@@ -15,10 +15,16 @@ public class RemoveProductCommand extends BaseCommandImp implements Command {
 	}
 	
 	@Override
-	public void execute() throws Exception {
-		int rowsCount = ProductDAO.removeProductsById(Integer.parseInt(params.get("id")));
-		if (0 == rowsCount) {
-			System.out.printf("There is nothing to delete");
+	public boolean execute() throws Exception {
+		boolean isQueryOK = false;
+		String paramsList[] = {"id"};
+		if (isParamsValid(paramsList)) {	
+			int rowsCount = ProductDAO.removeProductsById(Integer.parseInt(params.get("id")));
+			isQueryOK = true;
+			if (0 == rowsCount) {
+				System.out.println("There is nothing to delete");
+			}
 		}
+		return isQueryOK;
 	}
 }
